Fix bug in handling flags set to false
If the flag is set to false (e.g. no_preview = false in the project cli-config.js), then it was being added to the command anyways.
This escentially makes the true/false to mean add to the command vs not. We may consider changing the approach to not pass values, but instead just the list of flags to add.